Renamed the file according to its functionality
