+++ controllers/users/updateAvatar.js
